S-51801: No longer need to check if the user can edit a bug.
There were some calls being made in the RPC client that we've deemed no longer
necessary in the new REST client.

The call was checking to see if a the assigned to user could edit the
product that the bug was attached to.

From V1.pm we had this call

assign_user->can_edit_product($productid)

After looking at the capabilities of the new REST api, we discovered that
these types of validations were removed when moving from the RPC to the
RESTful form.

See this section for their documentation on what was removed from the
User:
https://wiki.mozilla.org/Bugzilla:API_Comparison#User

Acceptance tests that reflect our decision are
AT-52854 and AT-52853