S-51801: Upate a bug's field with a value.
The old version of Update (i.e. the Perl version) did two checks.
1) It checked if a field had been marked obsolete.
2) It checked if the field was of a particular type.

It did both of these things prior to executing an Update to the bug to
be persisted with things that had been configured via the ServiceHost
configuration client.

It did these every time the client ran and synced up betwee the Bugzilla
system and VersionOne.

Really, these checks are checks to make sure the user has not shot themself
in the foot and configured the system incorrectly. They don't need to
be run every single time an update occurs via this method.

After looking around at the RESTful API we discovered that we can easily
find out whether a field is of a particular type. See this page:
http://bugzilla.readthedocs.io/en/latest/api/core/v1/field.html

In our case we would be interested in the type Single value field I believe
and it would have a value of 2.

In the case a field is marked obsolete, we have to get more creative.
There really doesn't seem to be any indication via the RESTful Field API on the
Bugzilla side that shows if a field is marked obsolete.

But, if you post to a bug to update a field that was marked as obsolete,
you will get a HTTP 200 status code back (makes me kinda sad), but you will
also get a payload back that has what changes were made. If no changes were
applied, you will get back an empty changes element. If changes were applied
you will get back what fields changed and what their old and new values were.