S-51801: Create ChangeStatus overload to handle different cases and begin to split those cases out.
Rather than relying on special fields being there or not (which is almost
the same as passing in a boolean, another type of code smell), just simply
create versions of ChangeStatus that are used where they are needed.

This cleans up the code in each individual case, and you don't have to
carry as much context in your mind when reading it.