S-51801: Removing SkipCloseActions from the ReaderUpdater
As we looked at what this code was doing, it didn't make sense that it
needed to do this set of work.

When a Bug is created in Bugzilla, a corresponding Defect gets created in
VersionOne. Then, when the Defect is closed in VersionOne it get's updated
in Bugzilla to have it's custom field updated to 'Closed' and it's status
set appropriately (by default this would be a status of Resolved and Fixed).

The SkipCloseActions seemed to want to protect from making duplicate updates
to the Bug in Bugzilla if a user in VersionOne set a Defect to Closed, and
then Re-Opened it, made some changes, and then closed it again.

Oddly, it was also checking if the ReAssign user for State Changes that occur
on the VersionOne side was a valid user. That is a seperate concern from
any logic used in trying to decide if a Closed Defect/Bug should or should
not be processed. This added to the confusion of those involved.

This is seemingly overkill, and most probably not needed. After talking about it
as a team, we have chosen to remove this check for now. Should it need
to come back, we'll hopefully be able to track it back to here.