S-60148: Use the constructor of Bug to parse out from a JToken based set of data the values to populate the Bug class rather than having those properties messaged from the outside.
Also, make any of the properties read only that we can as they do not
have the need to be public, at least not all of them.

Move the DependsOn property from being a List of JTokens to being a collection
of int's. The rest of the system should not have to account for dealing with
JTokens, which is very much a parsing concern... not a domain concept.

Also, include tests for how a Bug parses out a known set of data that
currently gets returned from the Bugzilla API.